requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.18
joblib>=1.3.0
fastdtw>=0.3.4
stumpy>=1.12.0
xgboost>=1.7.0
pandas-ta
psutil>=5.9